Calculates ISS location

+++ structs.go
@@ -1,7 +1,7 @@
 package main
 
 type issCoords struct {
-	Latitude  float32 `json:"latitude"`
-	Longitude float32 `json:"longitude"`
-	Altitude  float32 `json:"altitude"`
+	Latitude  float64 `json:"latitude"`
+	Longitude float64 `json:"longitude"`
+	Altitude  float64 `json:"altitude"`
 }
